tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/146.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: SVG script adding src attribute </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>var t = async_test();</script>
<svg>
<script></script>
</svg>
<script>
t.step(function() {
  var s = document.querySelector("svg > script");
  s.src = "scripts/include-1.js";
});
onload = t.step_func(function() {
  // SVG <script> element uses href attribute, so src attribute is ignored.
  assert_array_equals(eventOrder, []);
  t.done();
});
</script>
